Add files via upload
I am almost done with the first fit algorithm working correctly in filling in-between the gaps of memory caused by freeing. I think the issue is when we split because our split I think assumes that we are at the end of the block but does not consider if we are in the middle of the block.